# FSE: Compute the block state of different entities separately