--- a/packages/block-editor/src/components/inner-blocks/index.native.js
+++ b/packages/block-editor/src/components/inner-blocks/index.native.js
@@ -1,204 +1,138 @@
 /**
- * External dependencies
- */
-import { pick, isEqual } from 'lodash';
-
-/**
  * WordPress dependencies
  */
-import { Component } from '@wordpress/element';
-import { withSelect, withDispatch } from '@wordpress/data';
-import {
-	synchronizeBlocksWithTemplate,
-	withBlockContentContext,
-} from '@wordpress/blocks';
-import isShallowEqual from '@wordpress/is-shallow-equal';
-import { compose } from '@wordpress/compose';
+import { useSelect } from '@wordpress/data';
+import { getBlockType, withBlockContentContext } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
 import ButtonBlockAppender from './button-block-appender';
 import DefaultBlockAppender from './default-block-appender';
+import useNestedSettingsUpdate from './use-nested-settings-update';
+import useInnerBlockTemplateSync from './use-inner-block-template-sync';
+import getBlockContext from './get-block-context';
+
+/**
+ * Internal dependencies
+ */
 import BlockList from '../block-list';
-import { withBlockEditContext } from '../block-edit/context';
+import { useBlockEditContext } from '../block-edit/context';
+import useBlockSync from '../provider/use-block-sync';
+import { BlockContextProvider } from '../block-context';
 
-class InnerBlocks extends Component {
-	constructor() {
-		super( ...arguments );
-		this.state = {
-			templateInProcess: !! this.props.template,
-		};
-		this.updateNestedSettings();
-	}
+/**
+ * InnerBlocks is a component which allows a single block to have multiple blocks
+ * as children. The UncontrolledInnerBlocks component is used whenever the inner
+ * blocks are not controlled by another entity. In other words, it is normally
+ * used for inner blocks in the post editor
+ *
+ * @param {Object} props The component props.
+ */
+function UncontrolledInnerBlocks( props ) {
+	const {
+		clientId,
+		allowedBlocks,
+		template,
+		templateLock,
+		templateInsertUpdatesSelection,
+		__experimentalMoverDirection,
+		renderAppender,
+		renderFooterAppender,
+		parentWidth,
+		horizontal,
+		contentResizeMode,
+		contentStyle,
+		onAddBlock,
+		onDeleteBlock,
+		marginVertical,
+		marginHorizontal,
+		horizontalAlignment,
+	} = props;
 
-	getTemplateLock() {
-		const { templateLock, parentLock } = this.props;
-		return templateLock === undefined ? parentLock : templateLock;
-	}
+	const block = useSelect( ( select ) =>
+		select( 'core/block-editor' ).getBlock( clientId )
+	);
 
-	componentDidMount() {
-		const { innerBlocks } = this.props.block;
-		// only synchronize innerBlocks with template if innerBlocks are empty or a locking all exists
-		if ( innerBlocks.length === 0 || this.getTemplateLock() === 'all' ) {
-			this.synchronizeBlocksWithTemplate();
-		}
+	useNestedSettingsUpdate( clientId, allowedBlocks, templateLock );
 
-		if ( this.state.templateInProcess ) {
-			this.setState( {
-				templateInProcess: false,
-			} );
-		}
-	}
+	useInnerBlockTemplateSync(
+		clientId,
+		template,
+		templateLock,
+		templateInsertUpdatesSelection
+	);
 
-	componentDidUpdate( prevProps ) {
-		const { template, block } = this.props;
-		const { innerBlocks } = block;
+	let blockList = (
+		<BlockList
+			marginVertical={ marginVertical }
+			marginHorizontal={ marginHorizontal }
+			rootClientId={ clientId }
+			renderAppender={ renderAppender }
+			renderFooterAppender={ renderFooterAppender }
+			withFooter={ false }
+			__experimentalMoverDirection={ __experimentalMoverDirection }
+			parentWidth={ parentWidth }
+			horizontalAlignment={ horizontalAlignment }
+			horizontal={ horizontal }
+			contentResizeMode={ contentResizeMode }
+			contentStyle={ contentStyle }
+			onAddBlock={ onAddBlock }
+			onDeleteBlock={ onDeleteBlock }
+		/>
+	);
 
-		this.updateNestedSettings();
-		// only synchronize innerBlocks with template if innerBlocks are empty or a locking all exists
-		if ( innerBlocks.length === 0 || this.getTemplateLock() === 'all' ) {
-			const hasTemplateChanged = ! isEqual(
-				template,
-				prevProps.template
-			);
-			if ( hasTemplateChanged ) {
-				this.synchronizeBlocksWithTemplate();
-			}
-		}
-	}
+	// Wrap context provider if (and only if) block has context to provide.
+	const blockType = getBlockType( block.name );
+	if ( blockType && blockType.providesContext ) {
+		const context = getBlockContext( block.attributes, blockType );
 
-	/**
-	 * Called on mount or when a mismatch exists between the templates and
-	 * inner blocks, synchronizes inner blocks with the template, replacing
-	 * current blocks.
-	 */
-	synchronizeBlocksWithTemplate() {
-		const { template, block, replaceInnerBlocks } = this.props;
-		const { innerBlocks } = block;
-
-		// Synchronize with templates. If the next set differs, replace.
-		const nextBlocks = synchronizeBlocksWithTemplate(
-			innerBlocks,
-			template
-		);
-		if ( ! isEqual( nextBlocks, innerBlocks ) ) {
-			replaceInnerBlocks( nextBlocks );
-		}
-	}
-
-	updateNestedSettings() {
-		const {
-			blockListSettings,
-			allowedBlocks,
-			updateNestedSettings,
-		} = this.props;
-
-		const newSettings = {
-			allowedBlocks,
-			templateLock: this.getTemplateLock(),
-		};
-
-		if ( ! isShallowEqual( blockListSettings, newSettings ) ) {
-			updateNestedSettings( newSettings );
-		}
-	}
-
-	render() {
-		const {
-			clientId,
-			renderAppender,
-			renderFooterAppender,
-			__experimentalMoverDirection,
-			parentWidth,
-			horizontal,
-			contentResizeMode,
-			contentStyle,
-			onAddBlock,
-			onDeleteBlock,
-			marginVertical,
-			marginHorizontal,
-			horizontalAlignment,
-		} = this.props;
-		const { templateInProcess } = this.state;
-
-		return (
-			<>
-				{ ! templateInProcess && (
-					<BlockList
-						marginVertical={ marginVertical }
-						marginHorizontal={ marginHorizontal }
-						rootClientId={ clientId }
-						renderAppender={ renderAppender }
-						renderFooterAppender={ renderFooterAppender }
-						withFooter={ false }
-						__experimentalMoverDirection={
-							__experimentalMoverDirection
-						}
-						parentWidth={ parentWidth }
-						horizontalAlignment={ horizontalAlignment }
-						horizontal={ horizontal }
-						contentResizeMode={ contentResizeMode }
-						contentStyle={ contentStyle }
-						onAddBlock={ onAddBlock }
-						onDeleteBlock={ onDeleteBlock }
-					/>
-				) }
-			</>
+		blockList = (
+			<BlockContextProvider value={ context }>
+				{ blockList }
+			</BlockContextProvider>
 		);
 	}
+
+	return blockList;
 }
 
-InnerBlocks = compose( [
-	withBlockEditContext( ( context ) => pick( context, [ 'clientId' ] ) ),
-	withSelect( ( select, ownProps ) => {
-		const {
-			isBlockSelected,
-			hasSelectedInnerBlock,
-			getBlock,
-			getBlockListSettings,
-			getBlockRootClientId,
-			getTemplateLock,
-		} = select( 'core/block-editor' );
-		const { clientId } = ownProps;
-		const block = getBlock( clientId );
-		const rootClientId = getBlockRootClientId( clientId );
+/**
+ * The controlled inner blocks component wraps the uncontrolled inner blocks
+ * component with the blockSync hook. This keeps the innerBlocks of the block in
+ * the block-editor store in sync with the blocks of the controlling entity. An
+ * example of an inner block controller is a template part block, which provides
+ * its own blocks from the template part entity data source.
+ *
+ * @param {Object} props The component props.
+ */
+function ControlledInnerBlocks( props ) {
+	useBlockSync( props );
+	return <UncontrolledInnerBlocks { ...props } />;
+}
 
-		return {
-			block,
-			blockListSettings: getBlockListSettings( clientId ),
-			hasOverlay:
-				block.name !== 'core/template' &&
-				! isBlockSelected( clientId ) &&
-				! hasSelectedInnerBlock( clientId, true ),
-			parentLock: getTemplateLock( rootClientId ),
-		};
-	} ),
-	withDispatch( ( dispatch, ownProps ) => {
-		const { replaceInnerBlocks, updateBlockListSettings } = dispatch(
-			'core/block-editor'
-		);
-		const {
-			block,
-			clientId,
-			templateInsertUpdatesSelection = true,
-		} = ownProps;
+/**
+ * Wrapped InnerBlocks component which detects whether to use the controlled or
+ * uncontrolled variations of the InnerBlocks component. This is the component
+ * which should be used throughout the application.
+ *
+ * @param {Object} props The component props.
+ */
+const InnerBlocks = ( props ) => {
+	const { clientId } = useBlockEditContext();
 
-		return {
-			replaceInnerBlocks( blocks ) {
-				replaceInnerBlocks(
-					clientId,
-					blocks,
-					block.innerBlocks.length === 0 &&
-						templateInsertUpdatesSelection
-				);
-			},
-			updateNestedSettings( settings ) {
-				dispatch( updateBlockListSettings( clientId, settings ) );
-			},
-		};
-	} ),
-] )( InnerBlocks );
+	const allProps = {
+		clientId,
+		...props,
+	};
+
+	// Detects if the InnerBlocks should be controlled by an incoming value.
+	return props.value && props.onChange ? (
+		<ControlledInnerBlocks { ...allProps } />
+	) : (
+		<UncontrolledInnerBlocks { ...allProps } />
+	);
+};
 
 // Expose default appender placeholders as components.
 InnerBlocks.DefaultBlockAppender = DefaultBlockAppender;