… docs/designers-developers/developers/data/data-core-block-editor.md
… packages/block-editor/CHANGELOG.md
… packages/block-editor/src/components/provider/index.js
… packages/block-editor/src/components/provider/index.native.js
… packages/block-editor/src/components/provider/test/use-block-sync.js
… packages/block-editor/src/components/provider/use-block-sync.js
… packages/block-editor/src/store/actions.js
… packages/block-editor/src/store/selectors.js
… packages/block-editor/src/store/test/reducer.js
… packages/block-editor/src/store/test/selectors.js
… packages/block-library/src/template-part/edit/inner-blocks.js
… packages/e2e-test-utils/README.md
… packages/e2e-test-utils/src/insert-block.js
… packages/e2e-tests/specs/experiments/multi-entity-editing.test.js
… packages/edit-site/src/components/add-template/index.js