## Description
Ultimately, this PR fixes dirty and undo state in the site editor related to different entities. Edits to template parts, nested template parts, and the root template are now considered separately. To test this, make sure that changes you make to one entity do not cause any other entity to become dirty. Additionally, test that undoing a parent entity does not delete a child entity from the editor. Also test that you do not dirty the editor by switching between different templates/template parts from the template selector dropdown.

A few aspects of the block-editor were changed to accomplish these changes.
1. This PR adds a bit of state which lets us keep track of which blocks are controlling their own inner blocks. This is important because a top-level entity, like a template in FSE, will often control all of the blocks in the editor. However, the template can render other entities, like a template part, which control and persist their own InnerBlocks. However, the template thinks that the template part's inner blocks are part of its own blocks, which becomes an issue when we look at the isDirty state. The template thinks that it is dirty even when only a template part has been updated.
2. We refactored the sync mechanism for the BlockEditorProvider so that it can also be used for inner block controllers. This means that the same code is used any time an entity and block-editor state need to stay in sync.
3. Various parts of the reducer state were fixed so that changes to one entity don't also affect others. For example, RESET_BLOCKS and REPLACE_INNER_BLOCKS were both deleting the blocks from other entities, which meant that those entities deleted their own content. Additionally, the cache state has been updated so that a parent block cache's are not updated past the controlling entity.
4. We also updated the `getBlock` selector to not return controlled inner blocks. This way, blocks which belong to a child entity are not returned when asking for the blocks of a parent entity. This also meant updating the `getBlocks` cache so that it does not refresh when blocks belonging to a different entity are changed.
5. We refactored InnerBlocks and BlockEditorProvider into functional components with hooks since we were already touching the code and needed to be able to use our new hook. This allows for more reusability between inner blocks and block editor provider, and even between inner blocks web and mobile component variants.

Many edge cases have been fixed so far in this PR, so it should be ready to go after we have more tests! Here is a list of a few things we were considering at the start of the PR that have now been fixed:
- [x] ~Not sure if nested controlled inner blocks work properly. My guess is that they do not work as expected.~ ✅ Nested inner block controllers should now work properly.
- [x] ~Will things like the undo stack still work correctly?~ ✅ Fixed
- [x] ~Need to probably update react native files~ ✅ Done
- [ ] Need to add some test cases so that entity dirty state behavior doesn't break in the future.

## How has this been tested?
Locally in edit-site, in the post/page editor, and via unit/e2e tests.

## Types of changes
Bug fix/enhancement

## Checklist:
- [ ] My code is tested.
- [x] My code follows the WordPress code style. <!-- Check code: `npm run lint`, Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/javascript/ -->
- [x] My code follows the accessibility standards. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/coding-standards/accessibility-coding-standards/ -->
- [x] My code has proper inline documentation. <!-- Guidelines: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/javascript/ -->
- [x] I've included developer documentation if appropriate. <!-- Handbook: https://developer.wordpress.org/block-editor/ -->
- [x] I've updated all React Native files affected by any refactorings/renamings in this PR. <!-- React Native mobile Gutenberg guidelines: https://github.com/WordPress/gutenberg/blob/master/docs/contributors/native-mobile.md -->