@mtias 2020-04-03
I am not entirely sure "controlled inner blocks" is the right abstraction for the motivation here. What other cases are to be considered here apart from template parts? Would reusable blocks be the same? How does "no knowledge over inner blocks of template parts" interact with:

- Undo / Redo stack.
- Blocks attempting to read the presence of blocks in the page across template parts boundaries. Should it be possible?
- Global state like block count, word count, etc.
- Traversing outline structure (getting information about heading levels across template part boundaries).

It might also be that conceptualizing it as a "provider" can lead to better nomenclature.