@noahtallen 2020-04-03
> What other cases are to be considered here apart from template parts

This is a great thought, though. Indeed, I can imagine some more scenarios that could work similarly to this one. For example, when a local block has set a direct templateLock on its own InnerBlocks, it does its own template validation separately from the global template validation. (see #11681 for more context). I think a very similar mechanism could be used to resolve that issue, so could our abstraction be clearer to allow more than just "controlledInnerBlocks"? Maybe we ought to be tracking InnerBlocks metadata, or something along those lines.

When it comes to reusable blocks, I'm not familiar with how those work. But, with some testing locally, I see that they do _not_ have the same issue we have with template parts. Editing a reusable block does not dirty the state of the entity which contains it. I'll need to figure out how reusable blocks manages to avoid the same problem, and maybe we can learn some lessons from it.

I wonder, since reusable blocks and template parts are so similar, if our UI and code around them should also work similarly