@noahtallen 2020-04-04
Looking into reusable blocks a bit, and I notice that they are implemented by:

1. Not including the reusable block sub-blocks in the main block editor provider. (which definitely affects the items mentioned above, like block count)
2. Rendering its own blocks in the editor as a little BlockEditorProvider with only the reusable blocks.

Do you think that we would move to a similar mechanism as the template part block and do away with the nested BlockEditorProvider?