@epiqueras 2020-04-22
Yeah, I suggested that we integrate #19741 into this, yesterday.

The issues discussed here happen because we clear the cache after resets, so we need to persist it. And, because edits bubble up to the root, when they should stop at the first parent without a `save` method, which we can easily do as shown in that PR.