@youknowriad 2020-04-23
@epiqueras 

> The issues discussed here happen because we clear the cache after resets, so we need to persist it.

I'm still having trouble understanding this :)