@epiqueras 2020-04-23
@youknowriad 

>>The issues discussed here happen because we clear the cache after resets, so we need to persist it.

>I'm still having trouble understanding this :)

Things like undoing will reset the block editor store blocks and wipe the cache, triggering an edit back to the editor store and dirtying the nested entities. We need to reuse caches for the blocks that haven't changed by referential equality after a reset to avoid this.

>I don't understand all the latest changes here, I feel these are just making things more complex.
Can you explain why setting the "controlled inner blocks" flag and just updating the "getBlocks" selector is not sufficient (and the cache key)? In theory it should be, so I need to understand better to know whether the added complexity is worth it.

I don't see how just that doing that could solve this issue. The block tree is immutable, so if you edit a block in a nested template part, all the blocks in its path to the root will be cloned. That means that even if you don't include the nested template part blocks in `getBlocks`, the equality check still won't pass.