@youknowriad 2020-04-23
> I don't see how just that doing that could solve this issue. The block tree is immutable, so if you edit a block in a nested template part, all the blocks in its path to the root will be cloned. 

The equality is check is based on the result of `getBlocks` so if the result of `getBlocks` don't change, no matter what the store contains (block-editor store and not core-data store) the equality check will be satisfied properly (getBlocks is only changed when the cache key changes).