@noahtallen 2020-05-05
I ran the performance test on this branch and compared it with master. TLDR is that after running each test a few times on each branch, I'm not really seeing any meaningful differences between the two. On some runs, master was faster, on some this branch was faster. Let me know if there is anything else to check for performance. I'm sure there are aspects of the code that can be further optimized!

This branch:
```
Loading Time:
Average time to load: 9291ms
Average time to DOM content load: 9257ms

Typing Performance:
Average time to type character: 37.15ms
Slowest time to type character: 84.56ms
Fastest time to type character: 34.17ms

Block Selection Performance:
Average time to select a block: 1204.94ms
Slowest time to select a block: 1441.1ms
Fastest time to select a block: 694.21ms
```

Master branch:
```
Loading Time:
Average time to load: 9262ms
Average time to DOM content load: 9229ms

Typing Performance:
Average time to type character: 36.31ms
Slowest time to type character: 85.88ms
Fastest time to type character: 32.87ms

Block Selection Performance:
Average time to select a block: 1213.41ms
Slowest time to select a block: 1372.58ms
Fastest time to select a block: 732.33ms
```